feat: extract package manager to a core module
